refactor: move map model classes back into the ui package
