+++ ground/src/main/java/com/google/android/ground/ui/home/HomeScreenViewModel.kt
@@ -22,7 +22,7 @@ import com.google.android.ground.model.Survey
 import com.google.android.ground.model.geometry.Point
 import com.google.android.ground.model.job.Job
 import com.google.android.ground.model.locationofinterest.LocationOfInterest
-import com.google.android.ground.model.map.MapLocationOfInterest
+import com.google.android.ground.ui.map.MapLocationOfInterest
 import com.google.android.ground.model.mutation.LocationOfInterestMutation
 import com.google.android.ground.model.mutation.Mutation
 import com.google.android.ground.repository.LocationOfInterestRepository
