+++ ground/src/main/java/com/google/android/ground/ui/home/mapcontainer/PolygonDrawingViewModel.kt
@@ -27,7 +27,7 @@ import com.google.android.ground.model.geometry.Point
 import com.google.android.ground.model.geometry.Polygon
 import com.google.android.ground.model.job.Job
 import com.google.android.ground.model.locationofinterest.LocationOfInterest
-import com.google.android.ground.model.map.MapLocationOfInterest
+import com.google.android.ground.ui.map.MapLocationOfInterest
 import com.google.android.ground.persistence.uuid.OfflineUuidGenerator
 import com.google.android.ground.rx.annotations.Hot
 import com.google.android.ground.system.LocationManager
