+++ ground/src/main/java/com/google/android/ground/ui/map/CameraPosition.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.model.map
+package com.google.android.ground.ui.map
 
 import com.google.android.ground.model.geometry.Coordinate
 import com.google.android.ground.model.geometry.Point
