+++ ground/src/main/java/com/google/android/ground/ui/map/MapFragment.kt
@@ -20,10 +20,6 @@ import androidx.annotation.IdRes
 import com.cocoahero.android.gmaps.addons.mapbox.MapBoxOfflineTileProvider
 import com.google.android.ground.model.geometry.Point
 import com.google.android.ground.model.locationofinterest.LocationOfInterest
-import com.google.android.ground.model.map.Bounds
-import com.google.android.ground.model.map.CameraPosition
-import com.google.android.ground.model.map.MapLocationOfInterest
-import com.google.android.ground.model.map.MapType
 import com.google.android.ground.rx.Nil
 import com.google.android.ground.rx.annotations.Hot
 import com.google.android.ground.ui.common.AbstractFragment
