+++ ground/src/main/java/com/google/android/ground/ui/map/MapType.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Google LLC
+ * Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.android.ground.model.map
+package com.google.android.ground.ui.map
 
-import com.google.android.ground.model.geometry.Coordinate
-import com.google.common.collect.ImmutableList
+import android.os.Parcelable
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+import kotlinx.parcelize.Parcelize
 
-interface Map {
-  val mapTypes: ImmutableList<MapType>
-  val viewport: Bounds
-  val zoom: Float
-  var mapType: MapType
-
-  fun moveCamera(coordinate: Coordinate)
-  fun moveCamera(coordinate: Coordinate, zoom: Float)
-  fun renderLocationOfInterest(locationOfInterest: MapLocationOfInterest)
-}
+/**
+ * MapType refers to the basemap shown below map LOIs and offline satellite imagery. It's called
+ * "map styles" in Mapbox and "basemaps" in Leaflet.
+ */
+@Parcelize
+data class MapType(
+  val type: Int,
+  @field:StringRes @param:StringRes val labelId: Int,
+  @field:DrawableRes @param:DrawableRes val imageId: Int
+) : Parcelable
