+++ ground/src/main/java/com/google/android/ground/ui/map/gms/GoogleMapsFragment.kt
@@ -38,15 +38,15 @@ import com.google.android.ground.model.geometry.*
 import com.google.android.ground.model.geometry.Polygon
 import com.google.android.ground.model.job.Style
 import com.google.android.ground.model.locationofinterest.LocationOfInterest
-import com.google.android.ground.model.map.Bounds
-import com.google.android.ground.model.map.MapLocationOfInterest
-import com.google.android.ground.model.map.MapType
+import com.google.android.ground.ui.map.Bounds
+import com.google.android.ground.ui.map.MapLocationOfInterest
+import com.google.android.ground.ui.map.MapType
 import com.google.android.ground.rx.Nil
 import com.google.android.ground.rx.annotations.Hot
 import com.google.android.ground.ui.MarkerIconFactory
 import com.google.android.ground.ui.common.AbstractFragment
 import com.google.android.ground.ui.map.*
-import com.google.android.ground.model.map.CameraPosition
+import com.google.android.ground.ui.map.CameraPosition
 import com.google.android.ground.ui.util.BitmapUtil
 import com.google.common.collect.ImmutableList
 import com.google.common.collect.ImmutableSet
